+++ lib/hyperglot/checks/check_mark_attachment.py
@@ -15,11 +15,12 @@ log.setLevel(logging.WARNING)
 
 class Check(CheckBase):
     """
-    Check to confirm mark attachment between base (and auxiliary) characters and marks.
+    Check to confirm mark attachment between base (and auxiliary) characters
+    and marks.
 
     - By default checks only base + mark.
-    - By default checks only unencoded base + mark combinations, but can be 
-        configured to also check decomposed combinations that do not exist 
+    - By default checks only unencoded base + mark combinations, but can be
+        configured to also check decomposed combinations that do not exist
         precomposed in the characters.
 
     Confirms input is indeed base letter and non-spacing mark. The check works
@@ -34,17 +35,41 @@ class Check(CheckBase):
     priority = 30
     logger = logging.getLogger("hyperglot.reporting.marks")
 
+    def precheck(
+        self,
+        orthography: Orthography,
+        checker: Checker,
+        **kwargs,
+    ) -> bool:
+        if not super().precheck(orthography, checker, **kwargs):
+            return False
+
+        if (
+            SupportLevel.AUX.value in self.options["check"]
+            and len(orthography.base_marks) == 0
+            and len(orthography.auxiliary_marks) == 0
+        ):
+            return False
+
+        if (
+            SupportLevel.BASE.value in self.options["check"]
+            and len(orthography.base_marks) == 0
+        ):
+            return False
+
+        return True
+
     def check(self, orthography: Orthography, checker: Checker, **kwargs) -> bool:
         """
         Check the mark attachment for the orthography.
         """
 
-        options = self._get_options(**kwargs)
+        super().check(orthography, checker, **kwargs)
 
         check_attachment = []
 
         chars = orthography.base
-        if SupportLevel.AUX.value in options["check"]:
+        if SupportLevel.AUX.value in self.options["check"]:
             chars.extend(orthography.auxiliary)
 
         # Mark positioning needs to at least work for all unencoded
@@ -53,7 +78,7 @@ class Check(CheckBase):
 
         # If checking against decomposed characters also base + mark combinations
         # that do not exist precomposed in the characters need to be checked.
-        if options["decomposed"]:
+        if self.options["decomposed"]:
             check_attachment.extend([c for c in chars if c not in checker.characters])
 
         missing_positioning = []